+++ README.md
@@ -54,7 +54,7 @@ OPENAI_MODEL: str = os.getenv('OPENAI_MODEL', 'gpt-3.5-turbo')  # Use to Open AI
 ## Endpoints
 
 - GET `/` - Get status(use to healthcheck)
-- POST `/chat` - Create a new messages, use to [Create chat completion](https://platform.openai.com/docs/api-reference/chat/create)
+- POST `/openai/chat` - Create a new messages, use to [Create chat completion](https://platform.openai.com/docs/api-reference/chat/create)
 
 ## Test
